Separate zoomable component
